Automatically set Game.type in setStatus
When setting/updating the authenticated user's status with the current game, the "Playing ...." info was not being updated in the discord client.

After some digging, looks like you have to specify a "type" on the status update in order to get it to show (This must be 0 or 1, Playing or Streaming, respectively).

This PR updates the docs to give some insight as to the requirement of that parameter, and also updates the `updateStatus` method to automatically populate the `type` property with 0 (Playing, from the Constants file) so it works out of the box a little more intutively.